Exercice 5 while resolved, include new text in array student on class StringText

+++ src/statics/DivisionStatic.java
@@ -2,7 +2,7 @@ package statics;
 
 public class DivisionStatic {
 
-	public static void averageAgeIndeterminateNumbers(int a, int b) {
+	public static void averageAgeIndeterminateNumbers(int a, double b) {
 
 		System.out.println("Average: " + String.format("%.2f", (double) b / a));
